Moved userId validation out of the public context
